Los "productos" (o puntos-negros) se van moviendo. No se ve muy bien porque estoy dibujando las cosas en el orden incorrecta (primero tendría que dibujar todos los elementos y después los productos para evitar problemas de superposición), pero ya se puede sentir el ritmo de Powerhouse sonando en la fábrica =P (tu tu tuuu turururu tu tuuu)
git-svn-id: http://svn2.xp-dev.com/svn/tp-tecnicas/trunk/tp-tecnicas@312 c5b30475-30d6-408d-831a-0400e7537d55

+++ src/controller/game/GamePanelController.java
@@ -200,7 +200,7 @@ public class GamePanelController {
 	}
 
 	private void repaintGroundPanel() {
-		double elapsedTickTime = this.timeCount / TICK_INTERVAL;
+		double elapsedTickTime = (double) this.timeCount / TICK_INTERVAL;
 		getGroundPanel().setElapsedTickTime(elapsedTickTime);
 		getGroundPanel().repaint();
 	}
